auto align but good forgot to commit
will add path planner event or smth later
